[frontend][refactor] Distinguish tags and values when no value
The case is:
```yaml
myTask:
  action: !com.example.foo
```
Previously this would report that either:
a) the tag is missing (obviously wrong)
b) the tag is unexpected (wrong), and the value is expected (correct)

Now it sees the tag properly and reports that the value is expected.

This is done by abstracting `YAMLValue` away with the `YamlValue` adapter hierarchy.