[compilation] Don't interleave stdout and stderr outputs from JIC
At the moment, the output of the Java compiler is returned in unstructured format: as a multi-line String.
This output is printed to the stdout/stderr of the JIC process.
Then this output is processed by two parallel threads (one for stdout, another for stderr) of the Amper process, and passed to the logger. Since the output is processed concurrently, lines from different messages, one of which is sent through stdout and another - through stderr, can interleave.

This commit encodes the "\n" character before sending it between processes, to make Amper process handling mechanism treat the whole compiler message as a single line.

A proper solution would be to serialize and send compiler messages in a structured format, this will be implemented later.