[frontend] Make generated paths in fragment accept build root
This way it's clearer from the API standpoint how they should be used (don't have to think what they are relative to—it's clear from the signature).

Also, now all the conventional paths don't depend on the actual contents of the Fragment source directories, but vice versa—the contents are populated by conventional paths.