[cli] Clean Java annotation processing directory before each build
Otherwise, if annotation processing is disabled after initially generating some sources they will stay there indefinitely. It is not a problem for the CLI as the generated directory isn't passed as another source for Java files, but it's the problem for the IDE that will resolve those sources (and they will never be deleted during builds).