[compilation] separate Java and Kotlin output
Kotlin and Java are compiled independently, Kotlin compilation goes first, and then Java compilation follows.
The Java compilation has an option to compile incrementally via the JPS. Sometimes, the JPS needs to clean the whole output directory to perform a full rebuild. But if Java and Kotlin compilations use the same output folder, there is no way for JPS to distinguish classes generated by the Kotlin and by the Java compiler. This commit fixes it.