[docs] Rename "DSL Reference" to just "Reference"
This was unnecessarily specific, and might bring confusion as to what
the "DSL" is in the context of Amper. It's not a domain-specific
language per se, just a configuration written in pure YAML.

Also, in the future, we might add other reference docs under this
same tab (for example, the CLI commands/options reference).