[cli] Do not crash downloads if a response doesn't have a Content-Length header
This integrity check is a "nice to have" to verify that everything is ok when we
can. If a response doesn't offer this possibility, it shouldn't completely
prevent the downloader from working.

Also, making this check optional allows us to use deflate/gzip compression for
some downloads where it could be useful.