[docs] sort files in the file tree, and modules in the project.yaml alphabetically
Files are usually sorted alphabetically on disk, so it is better to reflect the same state in the snippet.
Modules in the project.yaml better be sorted alphabetically as well, especially since it is consistent with the folders on disk.
"..." normally should be placed at the end of a list.