[cli] AMPER-4874 Ignore AMPER_JAVA_HOME if set to JBR21
Older versions of Amper set the AMPER_JAVA_HOME variable in
their script even when auto-provisioned. This means that
having AMPER_JAVA_HOME doesn't necessarily mean the user
decided to override it.
Because batch scripts forward their environment (including
local variables) to child processes, the update command
runs with this variable, and in turn passes it on to the
child process when running the new Amper version.

These old versions use the JBR 21, which is not compatible
with the new JRE 25 requirement. Therefore, we add this
special detection for it.