[docs] plugins-tutorial: adjust the initial section about the project structure
Previously, we were saying "this is the project structure we are aiming for", but we were not clearly communicating in the code, on which step the user should create the `app` module, on which step - the `build-config` module, etc. It is better to just create the expected project structure immediately.