[telemetry] Store telemetry logs in the 'telemetry' subfolder and rename it to amper_cli_traces.jsonl
We're going to have telemetry from child processes as well, so it would be more organized to group all of them in a subfolder. Renaming the telemetry file to "amper_main_traces.jsonl" also indicates that it is the telemetry of the main Amper process, rather than some of its children.