[frontend] Remove `Merged` trees stage.
The property of `Merged` trees was used very locally:
- Gutter icons in IDE
- Useless setting value
All of this functionality is currently disabled.

Regardless, it probably makes less sense to have `Merged` tree as a separate type in the type-system.
Its nodes are structurally the same as `Owned`.
If some feature needs this special property (different contexts are only at the "leaf" level) - it can do the processing itself,
using the single tree node type.

After this it's planned to have just two tree types: those with multiple contexts (multiple values per key) and those without.
So, "regular" trees and "refined" trees respectively (the naming may change in the future as refactoring goes).