[frontend][refactor] Type safe scalar nodes.
This refactoring does two things:
1. Makes `ScalarNode` a sealed interface with variants with type-safe value access.
2. Removes the usage of `TraceableValue` from the tree nodes entirely.

This simplifies work with the scalar nodes and makes it safer.

So now one mustn't remember to handle, e.g., `Enum`/`String`/`TraceableEnum` for enums.
Instead, there is `EnumNode` with all the necessary information available in a type-safe manner.
Same goes for other scalar types.

The `TraceableValue` wrapping only occurs when instantiating the `SchemaNode`s,
as `TraceableValue` is a thing specifically introduced to have traces for list/map elements there.

As a consequence, `TraceableValue` is now always guaranteed to have a consistent trace with the tree node it
 comes from.

Now, there is still the `isTraceableWrapped` flag on the `SchemaType`-level.
Its consistency still needs to be maintained for the `SchemaNode`s to be instantiated correctly.
We'll try to move away from that because it's unnecessary on the type system level.