[cli] Fix incorrect flow IDs in TeamCity format messages
Instead of setting the `teamCityFlowId` for the message, we ended up with setting the `uniqueId` which made test messages not being associated with the actual flows.

AmperTestFormatTest hasn't caught this due to the incorrect normalization. The normalization hasn't accumulated the mapping for flow IDs and instead used a clean map each time. This led to all messages having flow ID equal to '<normalized:0>' eliminating any grouping at all when doing the assertion.